+++ Changes.md
@@ -1,7 +1,17 @@
 1.6.x.x (relative to 1.6.6.1)
 =======
 
+Fixes
+-----
 
+- RenderPassEditor, Spreadsheet : Conformed menu buttons to match the rest of Gaffer by showing menu on button press, not release.
+- NodeEditor : Added missing widget for Color4fVectorDataPlug.
+
+API
+---
+
+- MenuButton : Added `set/getImmediate()` methods. These allow the button to execute a menu item immediately instead of showing the menu, if the menu has only a single item.
+- Signals : Added CatchingCombiner Python class, equivalent to the C++ template class.
 
 1.6.6.1 (relative to 1.6.6.0)
 =======